Add workflow example that demos the new waitpoint primitive
This PR adds a new example that enables creating audio summaries of newspaper articles.

This reference project shows a possible approach to implement workflows using Trigger.dev and ReactFlow.
It makes use of the Trigger.dev Realtime API and the new waitpoint primitive to implement a human-in-the-loop approach for approving the result of an AI workflow.